# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MagicAI  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/payment/": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get subscription plan of current user",
                "description": "Get subscription plan details of current user.",
                "operationId": "getCurrentPlan",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "No active subscription found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/payment/subscriptions/cancel-current": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Cancel current subscription plan of current user",
                "description": "Cancel current subscription plan details of current user.",
                "operationId": "cancelActiveSubscription",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/payment/plans/{plan_id?}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get all plans.",
                "description": "Get all plans.",
                "operationId": "plans",
                "parameters": [
                    {
                        "name": "plan_id",
                        "in": "path",
                        "description": "Id of plan to get details of.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/payment/orders/{order_id?}": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get all orders.",
                "description": "Get all orders. If order_id is provided, then it will return details of that order.",
                "operationId": "orders",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "description": "Id of order to get details of.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Permission Denied"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://127.0.0.1:8000/oauth/authorize",
                        "tokenUrl": "http://127.0.0.1:8000/oauth/token",
                        "refreshUrl": "http://127.0.0.1:8000/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}